Add files via upload
**A text based created while studying at the Tech Academy**

Gives a user the chance to be nice or mean to a dog that they encounter. 
You get a point valuation for each time you are mean or nice
after five points you are determined either a winner or a loser based on how nice you were to the dog! 

Never trust someone who doesn't like dogs!